@MitraDarja 2020-06-30
This calculation of minimisers is called robust winnowing, maybe we should mention that in the documentation and link to a paper, for example this one: https://www.biorxiv.org/content/10.1101/2020.02.11.943241v1
(Chapter 3 is the relevant here, I will probably present that paper soon in JC.)