@smehringer 2020-07-20
> LGTM, but do you want to add the robust winnowing in the documentation, so that a user can easily know about it?

Good point :D