@smehringer 2020-07-21
> What does the benchmark say (before/after) about non-repeat regions?

Good point. here the results
```
ungapped/50000/8/13       567911 ns       567898 ns         1181 Throughput[bp/s]=88.0317M/s
ungapped/50000/8/28       521850 ns       521719 ns         1321 Throughput[bp/s]=95.8235M/s
ungapped/50000/30/35      584811 ns       584798 ns         1168 Throughput[bp/s]=85.45M/s
ungapped/50000/30/50      519738 ns       519716 ns         1270 Throughput[bp/s]=96.1505M/s
gapped/50000/8/13        1209285 ns      1209246 ns          576 Throughput[bp/s]=41.3423M/s
gapped/50000/8/28        1128892 ns      1128831 ns          609 Throughput[bp/s]=44.2874M/s
gapped/50000/30/35       3003135 ns      3003134 ns          235 Throughput[bp/s]=16.6396M/s
gapped/50000/30/50       2835135 ns      2835062 ns          247 Throughput[bp/s]=17.6261M/s

ungapped/50000/8/13       664567 ns       664128 ns         1013 Throughput[bp/s]=75.2762M/s
ungapped/50000/8/28       585433 ns       585423 ns         1154 Throughput[bp/s]=85.3964M/s
ungapped/50000/30/35      683428 ns       683418 ns          996 Throughput[bp/s]=73.1193M/s
ungapped/50000/30/50      588969 ns       588574 ns         1150 Throughput[bp/s]=84.9019M/s
gapped/50000/8/13        1317202 ns      1317169 ns          530 Throughput[bp/s]=37.9549M/s
gapped/50000/8/28        1212461 ns      1212432 ns          574 Throughput[bp/s]=41.2337M/s
gapped/50000/30/35       3041196 ns      3041149 ns          226 Throughput[bp/s]=16.4316M/s
gapped/50000/30/50       2876608 ns      2876540 ns          242 Throughput[bp/s]=17.3719M/s
```

Hm, there seems to a be a small performance regression.. This is unexpected because the computation did not get more expensive I think. 
Could it be because the additional member variable make the whole iterator less light-weight?