@smehringer 2020-07-21
I can't increase the performance of non-repeat regions :( I don't know if a small performance regression on non-repeat regions is ok for a 2 fold improvement in repeat regions (plus better results! which I think is most important).