@eseiler 2020-07-21
> I can't increase the performance of non-repeat regions :( I don't know if a small performance regression on non-repeat regions is ok for a 2 fold improvement in repeat regions (plus better results! which I think is most important).

Yes, we definitely need this, we want robust winnowing almost always.

My profiler was inconclusive (can't to interruption based profiling on a virtual machine, so I would have to do it on a server).
I suspect it's the `std::distance`; even though they are random access iterators, we still have to do something and it's kinda the only thing we do that we didn't do before...
Would be nice to get the position out of `min_element` since it iterates over the range anyway :D

But in the profile I saw the deque reallocation taking quite some time, so we will probably get better performance once we have the ringbuffer.